+++ internal/pkg/dbmigrate/migrations/000095_notify_media_team_id.up.sql
@@ -1,2 +1,2 @@
-ALTER TABLE notify_medias ADD COLUMN team_id BIGINT NULL;
+ALTER TABLE notify_medias ADD COLUMN team_id BIGINT UNSIGNED NULL;
 ALTER TABLE notify_medias ADD INDEX idx_team_id (team_id);